Iniciada implementação de validação na Controller Account

+++ app/Repositories/Account.ts
@@ -1,27 +1,34 @@
 import AccountModel from 'App/Models/Account'
+import MovementModel from 'App/Models/Movement'
 
 export default class Account {
     public async index() {
-        return await AccountModel.query().preload('user').paginate(1,10)
+        return await AccountModel.query().preload('user').paginate(1, 10)
     }
 
-    public async store(data){
+    public async store(data) {
+        let validateNumberAccount = await AccountModel.query().where('number', data.number)
+
+        if(validateNumberAccount.length > 0){
+            return
+        }
+
         let newAccount = await AccountModel.create(data)
         await newAccount.load('user')
         return newAccount
     }
 
-    public async show(params){
+    public async show(params) {
         let { id } = params
         let account = await AccountModel.query().where('id', id).preload('user')
         return account
     }
 
-    public async update(request){
+    public async update(request) {
         let { id } = request.params()
         let account = await AccountModel.find(id)
 
-        if(!account){
+        if (!account) {
             return id
         }
 
@@ -37,8 +44,14 @@ export default class Account {
     public async destroy(params) {
         let account = await AccountModel.find(params.id)
 
-        if(!account){
-            return { error: 'Invalid ID', id: params.id }
+        if (!account) {
+            return { cod: 404, id: params.id }
+        }
+
+        let movements = await MovementModel.query().where('account_id', account.id)
+
+        if(movements && movements.length > 0){
+            return { cod: 304, id: params.id }
         }
 
         await account!.delete()
